+++ lib/src/list.dart
@@ -115,6 +115,7 @@ class MyCardList extends StatefulWidget {
   // 4. 卡片布局
   final double? cardHeight;
   final Widget Function(int)? cardSubtitle;
+  final Widget Function(int)? cardBelow;
   final double? leadingAndBodySpacing;
   final EdgeInsets? Function(int)? cardPadding;
   final EdgeInsets? Function(int)? cardMargin;
@@ -156,6 +157,7 @@ class MyCardList extends StatefulWidget {
     // 4. 卡片布局
     this.cardHeight,
     this.cardSubtitle,
+    this.cardBelow,
     this.leadingAndBodySpacing,
     this.cardPadding,
     this.cardMargin,
@@ -333,6 +335,7 @@ class MyCardListState extends State<MyCardList> {
           // 2. 布局和尺寸
           height: widget.cardHeight,
           subtitle: widget.cardSubtitle?.call(index),
+          below: widget.cardBelow?.call(index),
           leadingAndBodySpacing: widget.leadingAndBodySpacing,
           padding: widget.cardPadding?.call(index),
           margin: widget.cardMargin?.call(index),